# swagger/prime.yaml
swagger: '2.0'
info:
  contact:
    email: dp3@truss.works
  description: |
    The Prime API is a RESTful API that enables the Prime contractor to request information about upcoming moves, update the details and status of those moves, and make payment requests. It uses Mutual TLS for authentication procedures.

    All endpoints are located at `primelocal/prime/v1/`.

  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  title: Milmove Prime API
  version: 0.0.1
basePath: /prime/v1
schemes:
  - http
...
paths:
...
  '/mto-shipments':
    post:
      summary: createMTOShipment
      description: |
        Creates a MTO shipment for the specified Move Task Order.
        Required fields include:
        * Shipment Type
        * Customer requested pick-up date
        * Pick-up Address
        * Delivery Address
        * Releasing / Receiving agents

        Optional fields include:
        * Customer Remarks
        * Releasing / Receiving agents
        * An array of optional accessorial service item codes
      consumes:
        - application/json
      produces:
        - application/json
      operationId: createMTOShipment
      tags:
        - mtoShipment
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/CreateMTOShipment'
      responses:
        '200':
          description: Successfully created a MTO shipment.
          schema:
            $ref: '#/definitions/MTOShipment'
        '400':
          $ref: '#/responses/InvalidRequest'
        '404':
          $ref: '#/responses/NotFound'
        '422':
          $ref: '#/responses/UnprocessableEntity'
        '500':
          $ref: '#/responses/ServerError'
  '/mto-shipments/{mtoShipmentID}':
    put:
      summary: updateMTOShipment
      description: |
        Updates an existing shipment for a Move Task Order (MTO). Only the following fields can be updated using this endpoint:

        * `scheduledPickupDate`
        * `actualPickupDate`
        * `firstAvailableDeliveryDate`
        * `destinationAddress`
        * `pickupAddress`
        * `secondaryDeliveryAddress`
        * `secondaryPickupAddress`
        * `primeEstimatedWeight`
        * `primeActualWeight`
        * `shipmentType`
        * `agents` - all subfields except `mtoShipmentID`, `createdAt`, `updatedAt`. You cannot add new agents to a shipment.

        Note that some fields cannot be manually changed but will still be updated automatically, such as `primeEstimatedWeightRecordedDate` and `requiredDeliveryDate`.
      operationId: updateMTOShipment
      tags:
        - mtoShipment
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: mtoShipmentID
          description: UUID of the shipment being updated.
          required: true
          format: uuid
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/MTOShipment'
        - in: header
          name: If-Match
          type: string
          required: true
          description: >
            Optimistic locking is implemented via the `If-Match` header. If the ETag header does not match
            the value of the resource on the server, the server rejects the change with a `412 Precondition Failed` error.
      responses:
        '200':
          description: Successfully updated the MTO shipment.
          schema:
            $ref: '#/definitions/MTOShipment'
        '400':
          $ref: '#/responses/InvalidRequest'
        '401':
          $ref: '#/responses/PermissionDenied'
        '403':
          $ref: '#/responses/PermissionDenied'
        '404':
          $ref: '#/responses/NotFound'
        '412':
          $ref: '#/responses/PreconditionFailed'
        '422':
          $ref: '#/responses/UnprocessableEntity'
        '500':
          $ref: '#/responses/ServerError'
...
definitions:
  Address:
    type: object
    properties:
      id:
        type: string
        format: uuid
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
      streetAddress1:
        type: string
        example: 123 Main Ave
        title: Street address 1
      streetAddress2:
        type: string
        example: Apartment 9000
        x-nullable: true
        title: Street address 2
      streetAddress3:
        type: string
        example: Montmârtre
        x-nullable: true
        title: Address Line 3
      city:
        type: string
        example: Anytown
        title: City
      state:
        title: State
        type: string
        x-display-value:
          AL: AL
          AK: AK
          AR: AR
          AZ: AZ
          CA: CA
          CO: CO
          CT: CT
          DC: DC
          DE: DE
          FL: FL
          GA: GA
          HI: HI
          IA: IA
          ID: ID
          IL: IL
          IN: IN
          KS: KS
          KY: KY
          LA: LA
          MA: MA
          MD: MD
          ME: ME
          MI: MI
          MN: MN
          MO: MO
          MS: MS
          MT: MT
          NC: NC
          ND: ND
          NE: NE
          NH: NH
          NJ: NJ
          NM: NM
          NV: NV
          NY: NY
          OH: OH
          OK: OK
          OR: OR
          PA: PA
          RI: RI
          SC: SC
          SD: SD
          TN: TN
          TX: TX
          UT: UT
          VA: VA
          VT: VT
          WA: WA
          WI: WI
          WV: WV
          WY: WY
        enum:
          - AL
          - AK
          - AR
          - AZ
          - CA
          - CO
          - CT
          - DC
          - DE
          - FL
          - GA
          - HI
          - IA
          - ID
          - IL
          - IN
          - KS
          - KY
          - LA
          - MA
          - MD
          - ME
          - MI
          - MN
          - MO
          - MS
          - MT
          - NC
          - ND
          - NE
          - NH
          - NJ
          - NM
          - NV
          - NY
          - OH
          - OK
          - OR
          - PA
          - RI
          - SC
          - SD
          - TN
          - TX
          - UT
          - VA
          - VT
          - WA
          - WI
          - WV
          - WY
      postalCode:
        type: string
        format: zip
        title: ZIP
        example: '90210'
        pattern: '^(\d{5}([\-]\d{4})?)$'
      country:
        type: string
        title: Country
        x-nullable: true
        example: 'USA'
        default: USA
      eTag:
        type: string
        readOnly: true
    required:
      - streetAddress1
      - city
      - state
      - postalCode
  ClientError:
    type: object
    properties:
      title:
        type: string
      detail:
        type: string
      instance:
        type: string
        format: uuid
    required:
      - title
      - detail
      - instance
  CreateMTOShipment:
    type: object
    properties:
      moveTaskOrderID:
        example: 1f2270c7-7166-40ae-981e-b200ebdf3054
        format: uuid
        type: string
      shipmentType:
        $ref: '#/definitions/MTOShipmentType'
      requestedPickupDate:
        format: date
        type: string
        description: The date the customer requested that this shipment be picked up.
      customerRemarks:
        type: string
        example: handle with care
        x-nullable: true
      pickupAddress:
        $ref: '#/definitions/Address'
      destinationAddress:
        $ref: '#/definitions/Address'
      agents:
        $ref: '#/definitions/MTOAgents'
      pointOfContact:
        type: string
        description: Email or id of a contact person for this update
      mtoServiceItems:
        type: array
        items:
          $ref: '#/definitions/MTOServiceItem'
    required:
      - moveTaskOrderID
      - pickupAddress
      - destinationAddress
      - shipmentType
      - requestedPickupDate
...
  Error:
    properties:
      title:
        type: string
      detail:
        type: string
      instance:
        type: string
        format: uuid
    required:
      - title
      - detail
    type: object
...
  MTOAgent:
    properties:
      id:
        example: 1f2270c7-7166-40ae-981e-b200ebdf3054
        format: uuid
        readOnly: true
        type: string
      mtoShipmentID:
        example: 1f2270c7-7166-40ae-981e-b200ebdf3054
        format: uuid
        type: string
        readOnly: true
      createdAt:
        format: date-time
        type: string
        readOnly: true
      updatedAt:
        format: date-time
        type: string
        readOnly: true
      firstName:
        type: string
        x-nullable: true
      lastName:
        type: string
        x-nullable: true
      email:
        type: string
        format: x-email
        pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
        x-nullable: true
      phone:
        type: string
        format: telephone
        pattern: '^[2-9]\d{2}-\d{3}-\d{4}$'
        x-nullable: true
      agentType:
        $ref: '#/definitions/MTOAgentType'
    type: object
  MTOAgents:
    items:
      $ref: '#/definitions/MTOAgent'
    type: array
    maxItems: 2
  MTOAgentType:
    type: string
    title: MTO Agent Type
    example: RELEASING_AGENT
    enum:
      - RELEASING_AGENT
      - RECEIVING_AGENT
  MTOServiceItem:
    description: MTOServiceItem describes a base type of a service item. Polymorphic type. Both Move Task Orders and MTO Shipments will have MTO Service Items.
    type: object
    discriminator: modelType
    properties:
      id:
        example: 1f2270c7-7166-40ae-981e-b200ebdf3054
        format: uuid
        type: string
      moveTaskOrderID:
        example: 1f2270c7-7166-40ae-981e-b200ebdf3054
        format: uuid
        type: string
      mtoShipmentID:
        example: 1f2270c7-7166-40ae-981e-b200ebdf3054
        format: uuid
        type: string
      reServiceName:
        type: string
        readOnly: true
      status:
        $ref: '#/definitions/MTOServiceItemStatus'
      rejectionReason:
        example: item was too heavy
        type: string
        x-nullable: true
        readOnly: true
      modelType: # Base type and sub-types of MTOServiceItem
        $ref: '#/definitions/MTOServiceItemModelType'
      eTag:
        type: string
        readOnly: true
    required:
      - modelType
      - moveTaskOrderID
...
  MTOServiceItemModelType:
    description: >
      Describes all model sub-types for a MTOServiceItem model.
      Prime can only request the following service codes for which they will use the corresponding modelType
        * DOFSIT - MTOServiceItemDOFSIT
        * DDFSIT - MTOServiceItemDDFSIT
        * DOSHUT, DDSHUT - MTOServiceItemShuttle
        * DCRT, DCRTSA, DUCRT - MTOServiceItemDomesticCrating
    type: string
    enum:
      - MTOServiceItemBasic
      - MTOServiceItemDOFSIT
      - MTOServiceItemDDFSIT
      - MTOServiceItemShuttle
      - MTOServiceItemDomesticCrating
...
  MTOServiceItemStatus:
    description: Describes all statuses for a MTOServiceItem.
    type: string
    readOnly: true
    enum:
      - SUBMITTED
      - APPROVED
      - REJECTED
  MTOShipment:
    properties:
      moveTaskOrderID:
        example: 1f2270c7-7166-40ae-981e-b200ebdf3054
        format: uuid
        type: string
        readOnly: true
      id:
        example: 1f2270c7-7166-40ae-981e-b200ebdf3054
        format: uuid
        type: string
        readOnly: true
      createdAt:
        format: date-time
        type: string
        readOnly: true
      updatedAt:
        format: date-time
        type: string
        readOnly: true
      approvedDate:
        description: date when the shipment was given the status "APPROVED"
        format: date
        type: string
        readOnly: true
      scheduledPickupDate:
        format: date
        type: string
      requestedPickupDate:
        format: date
        type: string
        readOnly: true
      firstAvailableDeliveryDate:
        format: date
        type: string
      primeActualWeight:
        type: integer
        example: 4500
      primeEstimatedWeight:
        type: integer
        example: 4500
      primeEstimatedWeightRecordedDate:
        format: date
        type: string
        readOnly: true
      actualPickupDate:
        format: date
        type: string
      requiredDeliveryDate:
        format: date
        type: string
        readOnly: true
      agents:
        $ref: '#/definitions/MTOAgents'
      mtoServiceItems:
        items:
          $ref: '#/definitions/MTOServiceItem'
        type: array
      customerRemarks:
        type: string
        readOnly: true
        example: handle with care
        x-nullable: true
      shipmentType:
        $ref: '#/definitions/MTOShipmentType'
      pickupAddress:
        $ref: '#/definitions/Address'
      destinationAddress:
        $ref: '#/definitions/Address'
      secondaryPickupAddress:
        $ref: '#/definitions/Address'
      secondaryDeliveryAddress:
        $ref: '#/definitions/Address'
      eTag:
        type: string
        readOnly: true
      status:
        type: string
        readOnly: true
        enum:
          - APPROVED
          - SUBMITTED
          - REJECTED
      rejectionReason:
        type: string
        readOnly: true
        example: MTO Shipment not good enough
        x-nullable: true
      pointOfContact:
        type: string
        description: Email or id of a contact person for this update.
...
  MTOShipmentType:
    type: string
    title: Shipment Type
    example: HHG
    enum:
      - HHG
      - INTERNATIONAL_HHG
      - INTERNATIONAL_UB
    x-display-value:
      HHG: HHG
      INTERNATIONAL_HHG: International HHG
      INTERNATIONAL_UB: International UB
...
  ValidationError:
    allOf:
      - $ref: '#/definitions/ClientError'
      - type: object
    properties:
      invalidFields:
        type: object
        additionalProperties:
          description: List of errors for the field
          type: array
          items:
            type: string
    required:
      - invalidFields
responses:
  InvalidRequest:
    description: The request payload is invalid.
    schema:
      $ref: '#/definitions/ClientError'
  NotFound:
    description: The requested resource wasn't found.
    schema:
      $ref: '#/definitions/ClientError'
...
  PermissionDenied:
    description: The request was denied.
    schema:
      $ref: '#/definitions/ClientError'
  ServerError:
    description: A server error occurred.
    schema:
      $ref: '#/definitions/Error'
...
  PreconditionFailed:
    description: Precondition failed, likely due to a stale eTag (If-Match). Fetch the request again to get the updated eTag value.
    schema:
      $ref: '#/definitions/ClientError'
  UnprocessableEntity:
    description: The payload was unprocessable.
    schema:
      $ref: '#/definitions/ValidationError'